Move cookie code to new CookieController class
Access restoreLoginCookies from CookieController
Access setCookieHeaders from CookieController
Access clearLoginCookies from CookieController
Access logCookies from CookieController
Access saveLoginCookies from CookieController
Access getCookieString from CookieController
