+++ Awful.apk/src/main/java/com/ferg/awfulapp/AwfulLoginActivity.java
@@ -46,6 +46,7 @@ import android.widget.Toast;
 
 import com.android.volley.NetworkResponse;
 import com.android.volley.VolleyError;
+import com.ferg.awfulapp.network.CookieController;
 import com.ferg.awfulapp.network.NetworkUtils;
 import com.ferg.awfulapp.preferences.AwfulPreferences;
 import com.ferg.awfulapp.preferences.Keys;
@@ -71,7 +72,7 @@ public class AwfulLoginActivity extends AwfulActivity {
 
         setContentView(R.layout.login);
 
-        NetworkUtils.logCookies();
+        CookieController.logCookies();
 
         mLogin = (Button) findViewById(R.id.login);
         mUsername = (EditText) findViewById(R.id.username);
@@ -162,7 +163,7 @@ public class AwfulLoginActivity extends AwfulActivity {
                     }
                 }
                 if (response != null && response.statusCode == HttpStatus.SC_MOVED_TEMPORARILY) {
-                    Boolean result = NetworkUtils.saveLoginCookies(getApplicationContext());
+                    Boolean result = CookieController.saveLoginCookies(getApplicationContext());
                     if(result){
                         // TODO: this should probably be handled by firing a ProfileRequest and getting the username from there, maybe through SyncManager
                         AwfulPreferences prefs = AwfulPreferences.getInstance(getApplicationContext());
